+++ src/backend/compare/report.ts
@@ -2,7 +2,8 @@ import { resolve } from 'node:path';
 import { existsSync, readFileSync, unlinkSync, rmSync } from 'node:fs';
 import { execFile, ChildProcessPromise } from 'promisify-child-process';
 
-import { Database, DatabaseConfig } from '../../db.js';
+import type { DatabaseConfig } from '../db/types.js';
+import { Database } from '../db/db.js';
 import { assert, log } from '../../logging.js';
 import { completeRequest, startRequest } from '../../perf-tracker.js';
 import { robustPath, siteConfig } from '../../util.js';
