+++ src/backend/rebench/results.ts
@@ -1,9 +1,9 @@
 import { ParameterizedContext } from 'koa';
 import { ValidateFunction } from 'ajv';
 
-import { BenchmarkData } from '../../api.js';
-import { Database } from '../../db.js';
-import { createValidator } from '../../api-validator.js';
+import { BenchmarkData } from '../../shared/api.js';
+import { Database } from '../db/db.js';
+import { createValidator } from './api-validator.js';
 import { DEBUG } from '../../util.js';
 import { log } from '../../logging.js';
 import { completeRequest, startRequest } from '../../perf-tracker.js';
