+++ src/timeline-calc.ts
@@ -1,4 +1,4 @@
-import { Database } from './db.js';
+import { Database } from './backend/db/db.js';
 import { SummaryStatistics } from './stats.js';
 import { robustSrcPath } from './util.js';
 import { Worker } from 'node:worker_threads';
