+++ src/views/plots.ts
@@ -1,5 +1,5 @@
-import type { AllResults, PlotData, TimelineResponse } from 'api';
-import type { Source } from 'db.js';
+import type { AllResults, PlotData, TimelineResponse } from '../shared/api.js';
+import type { Source } from '../backend/db/types.js';
 import { filterCommitMessage } from './render.js';
 import uPlot from '/static/uPlot.esm.min.js';
 import type { WarmupData, WarmupDataPerCriterion } from './view-types.js';
