+++ tests/backend/compare/prep-data.test.ts
@@ -3,13 +3,13 @@ import { describe, expect, it } from '@jest/globals';
 import { readFileSync, mkdirSync, existsSync } from 'node:fs';
 
 import { robustPath } from '../../../src/util.js';
-import {
+import type {
   MeasurementData,
   Measurements,
   ProcessedResult,
   RevisionComparison,
   RunSettings
-} from '../../../src/db.js';
+} from '../../../src/backend/db/types.js';
 import {
   ResultsByBenchmark,
   ResultsByExeSuiteBenchmark,
@@ -26,7 +26,7 @@ import {
   prepareCompareView
 } from '../../../src/backend/compare/prep-data.js';
 import { ComparisonStatsWithUnit } from '../../../src/stats.js';
-import { collateMeasurements } from '../../../src/db-data-processing.js';
+import { collateMeasurements } from '../../../src/backend/compare/db-data.js';
 import {
   ByExeSuiteComparison,
   CompareStatsRow,
@@ -34,7 +34,7 @@ import {
   MissingData
 } from '../../../src/views/view-types.js';
 import { prepareTemplate } from '../../../src/templates.js';
-import * as dataFormatters from '../../../src/data-format.js';
+import * as dataFormatters from '../../../src/shared/data-format.js';
 import * as viewHelpers from '../../../src/views/helpers.js';
 import {
   initJestMatchers,
