+++ tests/cache-validity.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it } from '@jest/globals';
-import { TimedCacheValidity } from '../src/db.js';
+import { TimedCacheValidity } from '../src/backend/db/timed-cache-validity.js';
 
 async function delayOf(ms): Promise<void> {
   return new Promise((resolve) => {
