+++ tests/report.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, afterAll, it } from '@jest/globals';
 import { readFileSync, unlinkSync, rmSync, existsSync } from 'fs';
-import { DatabaseConfig } from '../src/db.js';
+import type { DatabaseConfig } from '../src/backend/db/types.js';
 import { getDirname } from '../src/util.js';
 import {
   getOutputImageFolder,
